Tests/LibWeb/Text/expected/css/transform-skew-identity-interpolation.txt

matrix(1, 0, 0.57735, 1, 0, 0)
